+++ mobile/android/android-components/components/compose/awesomebar/src/main/java/mozilla/components/compose/browser/awesomebar/AwesomeBar.kt
@@ -96,7 +96,7 @@ fun AwesomeBar(
 
         // This state does not need to be remembered, because it can change if the providers list changes.
         @SuppressLint("UnrememberedMutableState")
-        val suggestions = derivedStateOf { fetcher.state.value }.value.toSortedMap(compareBy { it.title })
+        val suggestions = derivedStateOf { fetcher.state.value }.value.toSortedMap(compareByDescending { it.priority })
 
         LaunchedEffect(text, fetcher) {
             fetcher.fetch(text)