+++ mobile/android/android-components/components/compose/awesomebar/src/test/java/mozilla/components/compose/browser/awesomebar/internal/SuggestionFetcherTest.kt
@@ -6,12 +6,14 @@ package mozilla.components.compose.browser.awesomebar.internal
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.Job
+import mozilla.components.concept.awesomebar.AwesomeBar
 import mozilla.components.concept.awesomebar.AwesomeBar.SuggestionProvider
 import mozilla.components.concept.awesomebar.AwesomeBar.SuggestionProviderGroup
 import mozilla.components.support.test.any
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.rule.MainCoroutineRule
 import mozilla.components.support.test.rule.runTestOnMain
+import org.junit.Assert.assertEquals
 import org.junit.Rule
 import org.junit.Test
 import org.mockito.Mockito.doAnswer
@@ -37,5 +39,31 @@ class SuggestionFetcherTest {
 
         orderVerifier.verify(previousFetchJob)!!.cancel()
         orderVerifier.verify(fetcher).fetchFrom(providerGroup, provider, "test", null)
+    }
+
+    @Test
+    fun `GIVEN a suggestion group THEN the group's priority becomes highest suggestions' score within the group`() = runTestOnMain {
+        val provider: SuggestionProvider = mock()
+        val providerGroup = SuggestionProviderGroup(listOf(provider))
+        val suggestions = listOf(
+            AwesomeBar.Suggestion(
+                provider = provider,
+                score = Int.MAX_VALUE,
+            ),
+            AwesomeBar.Suggestion(
+                provider = provider,
+                score = Int.MIN_VALUE,
+            ),
+        )
+        val fetcher = spy(SuggestionFetcher(listOf(providerGroup), null))
+
+        fetcher.processResultFrom(
+            group = providerGroup,
+            provider = provider,
+            suggestions = suggestions,
+            profilerStartTime = null,
+        )
+
+        assertEquals(providerGroup.priority, Int.MAX_VALUE)
     }
 }